requirements/local.txt
# Local development dependencies go here
-r base.txt
coverage==3.6
django-discover-runner==0.4
Sphinx==1.2b1
fabtools==0.17.0
django-debug-toolbar==1.2.1